AC-848: converge CLI agent-task path onto task_runner's guardrail-adjusted threshold
_run_agent_task in cli.py drove generation lifecycle state directly and
never applied the guardrail-adjusted effective_met_threshold logic that
execution/task_runner.py::_process_task applies (objective/evaluator
guardrail payloads gating the raw met_threshold). The two paths could
report different met_threshold for identical inputs.

Extract TaskConfig and the payload/guardrail builder functions out of
task_runner.py into a new shared module,
execution/agent_task_completion.py (logic unchanged, calls renamed
1:1), plus a new compute_effective_met_threshold() that factors out the
previously-duplicated boolean gate. task_runner.py now imports from the
shared module instead of defining locally. cli.py's _run_agent_task
builds a minimal TaskConfig from AppSettings judge fields and applies
the same evaluator-guardrail gate after the loop finishes, so both
paths now share one implementation.
